replace iterators
